[Misc] Move create_kv_caches to kv_cache_manager
