+++ tests/models/jax/test_qwen2.py
@@ -11,7 +11,7 @@ from vllm.config import ModelConfig
 
 from tpu_commons.models.jax.attention_metadata import AttentionMetadata
 from tpu_commons.models.jax.qwen2 import Qwen2ForCausalLM
-from tpu_commons.runner import utils as runner_utils
+from tpu_commons.runner.jax.kv_cache_manager import create_kv_caches
 
 
 class MockVllmConfig:
@@ -128,7 +128,7 @@ class TestQwen2ForCausalLM:
         model.load_weights(rng)
 
         # Test model forward
-        kv_caches = runner_utils.create_kv_caches(
+        kv_caches = create_kv_caches(
             num_blocks=4,
             block_size=32,
             num_kv_heads=num_kv_heads,
