+++ tests/runner/test_utils.py
@@ -9,48 +9,13 @@ import jax.numpy as jnp
 import numpy as np
 import pytest
 from jax._src.interpreters import pxla
-from jax.sharding import Mesh, NamedSharding, PartitionSpec
 
-from tpu_commons.models.jax.attention import get_kv_cache_shape_with_mesh
 from tpu_commons.runner.utils import (
     PHASED_PROFILER_NUM_STEPS_TO_PROFILE_FOR, ForbidCompile, InferencePhase,
-    LatencyTracker, PhasedBasedProfiler, create_kv_caches,
-    determine_do_sampling, determine_phase_from_batch_composition_stats,
-    get_batch_composition_stats, get_padded_num_reqs_with_upper_limit,
-    get_padded_token_len, get_req_paddings, get_token_paddings,
-    pad_to_multiple)
-
-
-def test_determine_do_sampling():
-    """Tests the determine_do_sampling function."""
-    assert not determine_do_sampling(top_k=50, temperature=0.0)
-    assert not determine_do_sampling(top_k=1, temperature=0.7)
-    assert not determine_do_sampling(top_k=1, temperature=0.0)
-    assert determine_do_sampling(top_k=10, temperature=0.5)
-
-
-def test_pad_to_multiple():
-    """Tests the pad_to_multiple function."""
-    assert pad_to_multiple(16, 8) == 16
-    assert pad_to_multiple(17, 8) == 24
-    assert pad_to_multiple(24, 8) == 24
-    assert pad_to_multiple(1, 8) == 8
-
-    # Test with max_limit
-    assert pad_to_multiple(17, 8, max_limit=20) == 20
-    assert pad_to_multiple(17, 8, max_limit=30) == 24
-    assert pad_to_multiple(25, 8, max_limit=30) == 30
-
-    # Test with keep_one
-    assert pad_to_multiple(1, 8, keep_one=True) == 1
-    assert pad_to_multiple(2, 8, keep_one=True) == 8
-    assert pad_to_multiple(1, 8, keep_one=False) == 8
-
-    # Test invalid inputs
-    with pytest.raises(AssertionError):
-        pad_to_multiple(0, 8)
-    with pytest.raises(AssertionError):
-        pad_to_multiple(-5, 8)
+    LatencyTracker, PhasedBasedProfiler,
+    determine_phase_from_batch_composition_stats, get_batch_composition_stats,
+    get_padded_num_reqs_with_upper_limit, get_padded_token_len,
+    get_req_paddings, get_token_paddings)
 
 
 def test_get_padded_num_reqs_with_upper_limit():
@@ -243,57 +208,6 @@ def test_forbid_compile_raises_on_new_shape(jitted_function, jnp_array_input,
     with pytest.raises(RuntimeError, match=expected_error_message):
         with ForbidCompile(message=expected_error_message):
             jitted_function(jnp_array_input_new)
-
-
-@pytest.fixture
-def mesh():
-    devices = jax.devices()
-    return Mesh(devices, axis_names=("model", ))
-
-
-def test_create_kv_caches(mesh: Mesh):
-    """
-    Tests that `create_kv_caches` correctly allocates and shards the KV caches
-    for all specified layers.
-    """
-    num_blocks = 64
-    block_size = 16
-    num_kv_heads = 8
-    head_size = 128
-    layer_names = ["decoder.0", "decoder.1", "decoder.2"]  # Test with 3 layers
-    devices = jax.devices()
-
-    expected_sharding = NamedSharding(mesh, PartitionSpec())
-    expected_dtype = jnp.bfloat16
-    expected_shape = get_kv_cache_shape_with_mesh(mesh, num_blocks, block_size,
-                                                  num_kv_heads, head_size,
-                                                  expected_dtype)
-
-    with patch("tpu_commons.logger.init_logger",
-               return_value=MagicMock()), patch(
-                   "tpu_commons.runner.utils.utils.hbm_usage_gb",
-                   return_value=[(0.0, 0.0), (0.0, 0.0)]):
-        kv_caches = create_kv_caches(
-            num_blocks=num_blocks,
-            block_size=block_size,
-            num_kv_heads=num_kv_heads,
-            head_size=head_size,
-            mesh=mesh,
-            layer_names=layer_names,
-            devices=devices,
-        )
-
-        assert isinstance(kv_caches, list)
-        assert len(kv_caches) == len(layer_names)
-
-        for cache_array in kv_caches:
-            assert isinstance(cache_array, jax.Array)
-            assert cache_array.shape == expected_shape
-            assert cache_array.dtype == expected_dtype
-            assert cache_array.sharding == expected_sharding
-
-        # Ensure that separate array objects were created for each layer
-        assert kv_caches[0] is not kv_caches[1]
 
 
 class MockInputBatch:
